Separate widget tests in different files
