+++ test/test_widget_textinput.py
@@ -1,0 +1,667 @@
+"""
+pygame-menu
+https://github.com/ppizarror/pygame-menu
+
+TEST WIDGET - TEXTINPUT
+Test TextInput and ColorInput widgets.
+
+License:
+-------------------------------------------------------------------------------
+The MIT License (MIT)
+Copyright 2017-2021 Pablo Pizarro R. @ppizarror
+
+Permission is hereby granted, free of charge, to any person obtaining a
+copy of this software and associated documentation files (the "Software"),
+to deal in the Software without restriction, including without limitation
+the rights to use, copy, modify, merge, publish, distribute, sublicense,
+and/or sell copies of the Software, and to permit persons to whom the Software
+is furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
+WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+-------------------------------------------------------------------------------
+"""
+
+__all__ = ['TextInputWidgetTest']
+
+from test._utils import MenuUtils, surface, PygameEventUtils, TEST_THEME, PYGAME_V2, \
+    SYS_PLATFORM_OSX, BaseTest
+
+import pygame
+import pygame_menu
+import pygame_menu.controls as ctrl
+
+from pygame_menu.widgets.core.widget import WidgetTransformationNotImplemented
+
+
+class TextInputWidgetTest(BaseTest):
+
+    # noinspection SpellCheckingInspection,PyTypeChecker
+    def test_textinput(self) -> None:
+        """
+        Test TextInput widget.
+        """
+        if SYS_PLATFORM_OSX:
+            return
+
+        menu = MenuUtils.generic_menu()
+
+        # Assert bad settings
+        self.assertRaises(ValueError,
+                          lambda: menu.add.text_input('title',
+                                                      input_type=pygame_menu.locals.INPUT_FLOAT,
+                                                      default='bad'))
+        self.assertRaises(ValueError,  # Default and password cannot coexist
+                          lambda: menu.add.text_input('title',
+                                                      password=True,
+                                                      default='bad'))
+
+        # Create text input widget
+        textinput = menu.add.text_input('title', input_underline='_')
+        textinput.set_value('new_value')  # No error
+        textinput._selected = False
+        textinput.draw(surface)
+        textinput.select(update_menu=True)
+        textinput.draw(surface)
+        self.assertEqual(textinput.get_value(), 'new_value')
+        textinput.clear()
+        self.assertEqual(textinput.get_value(), '')
+
+        password_input = menu.add.text_input('title', password=True, input_underline='_')
+        self.assertRaises(ValueError,  # Password cannot be set
+                          lambda: password_input.set_value('new_value'))
+        password_input.set_value('')  # No error
+        password_input._selected = False
+        password_input.draw(surface)
+        password_input.select(update_menu=True)
+        password_input.draw(surface)
+        self.assertEqual(password_input.get_value(), '')
+        password_input.clear()
+        self.assertEqual(password_input.get_value(), '')
+
+        # Create selection box
+        string = 'the text'
+        textinput._cursor_render = True
+        textinput.set_value(string)
+        textinput._select_all()
+        self.assertEqual(textinput._get_selected_text(), 'the text')
+        textinput.draw(surface)
+        textinput._unselect_text()
+        textinput.draw(surface)
+
+        # Test maxchar and undo/redo
+        textinput = menu.add.text_input('title',
+                                        input_underline='_',
+                                        maxchar=20)
+        textinput.set_value('the size of this textinput is way greater than the limit')
+        self.assertEqual(textinput.get_value(), 'eater than the limit')  # same as maxchar
+        self.assertEqual(textinput._cursor_position, 20)
+        textinput._undo()  # This must set default at ''
+        self.assertEqual(textinput.get_value(), '')
+        textinput._redo()
+        self.assertEqual(textinput.get_value(), 'eater than the limit')
+        textinput.draw(surface)
+        textinput._copy()
+        textinput._paste()
+        textinput._block_copy_paste = False
+        textinput._select_all()
+        textinput._cut()
+        self.assertEqual(textinput.get_value(), '')
+        textinput._undo()
+        self.assertEqual(textinput.get_value(), 'eater than the limit')
+
+        # Test copy/paste
+        textinput_nocopy = menu.add.text_input('title',
+                                               input_underline='_',
+                                               maxwidth=20,
+                                               copy_paste_enable=False)
+        textinput_nocopy.set_value('this cannot be copied')
+        textinput_nocopy._copy()
+        textinput_nocopy._paste()
+        textinput_nocopy._cut()
+        self.assertEqual(textinput_nocopy.get_value(), 'this cannot be copied')
+
+        # Test copy/paste without block
+        textinput_copy = menu.add.text_input('title',
+                                             input_underline='_',
+                                             maxwidth=20,
+                                             maxchar=20)
+        textinput_copy.set_value('this value should be cropped as this is longer than the max char')
+        self.assertFalse(textinput_copy._block_copy_paste)
+        textinput_copy._copy()
+        self.assertTrue(textinput_copy._block_copy_paste)
+        textinput_copy._block_copy_paste = False
+        textinput_copy._select_all()
+        textinput_copy._cut()
+        self.assertEqual(textinput_copy.get_value(), '')
+        textinput_copy._block_copy_paste = False
+        textinput_copy._paste()
+        #  self.assertEqual(textinput_copy.get_value(), 'er than the max char')
+        textinput_copy._cut()
+        textinput_copy._block_copy_paste = False
+        # self.assertEqual(textinput_copy.get_value(), '')
+        textinput_copy._valid_chars = ['e', 'r']
+        textinput_copy._paste()
+        # noinspection SpellCheckingInspection
+        # self.assertEqual(textinput_copy.get_value(), 'erer')
+
+        # Assert events
+        textinput.update(PygameEventUtils.key(0, keydown=True, testmode=False))
+        PygameEventUtils.test_widget_key_press(textinput)
+        textinput.update(PygameEventUtils.key(ctrl.KEY_APPLY, keydown=True))
+        textinput.update(PygameEventUtils.key(pygame.K_LSHIFT, keydown=True))
+        textinput.clear()
+
+        # Type
+        textinput.update(PygameEventUtils.key(pygame.K_t, keydown=True, char='t'))
+        textinput.update(PygameEventUtils.key(pygame.K_e, keydown=True, char='e'))
+        textinput.update(PygameEventUtils.key(pygame.K_s, keydown=True, char='s'))
+        textinput.update(PygameEventUtils.key(pygame.K_t, keydown=True, char='t'))
+
+        # Keyup
+        textinput.update(PygameEventUtils.key(pygame.K_a, keyup=True, char='a'))
+        self.assertEqual(textinput.get_value(), 'test')  # The text we typed
+
+        # Ctrl events
+        textinput.update(PygameEventUtils.keydown_mod_ctrl(pygame.K_c))  # copy
+        textinput.update(PygameEventUtils.keydown_mod_ctrl(pygame.K_v))  # paste
+        textinput.update(PygameEventUtils.keydown_mod_ctrl(pygame.K_z))  # undo
+        self.assertEqual(textinput.get_value(), 'tes')
+        textinput.update(PygameEventUtils.keydown_mod_ctrl(pygame.K_y))  # redo
+        self.assertEqual(textinput.get_value(), 'test')
+        textinput._select_all()
+        textinput.update(PygameEventUtils.keydown_mod_ctrl(pygame.K_x))  # cut
+        self.assertEqual(textinput.get_value(), '')
+        textinput.update(PygameEventUtils.keydown_mod_ctrl(pygame.K_z))  # undo
+        self.assertEqual(textinput.get_value(), 'test')
+        textinput.update(PygameEventUtils.keydown_mod_ctrl(pygame.K_y))  # redo
+        self.assertEqual(textinput.get_value(), '')
+        textinput.update(PygameEventUtils.keydown_mod_ctrl(pygame.K_z))  # undo
+        self.assertEqual(textinput.get_value(), 'test')
+
+        # Test selection, if user selects all and types anything the selected
+        # text must be destroyed
+        textinput.update(PygameEventUtils.keydown_mod_ctrl(pygame.K_a))  # select all
+        textinput._unselect_text()
+        self.assertEqual(textinput._get_selected_text(), '')
+        textinput._select_all()
+        self.assertEqual(textinput._get_selected_text(), 'test')
+        textinput._unselect_text()
+        self.assertEqual(textinput._get_selected_text(), '')
+        textinput.update(PygameEventUtils.keydown_mod_ctrl(pygame.K_a))
+        self.assertEqual(textinput._get_selected_text(), 'test')
+        textinput.update(PygameEventUtils.key(pygame.K_t, keydown=True, char='t'))
+        textinput.update(PygameEventUtils.key(pygame.K_ESCAPE, keydown=True))
+
+        # Now the value must be t
+        self.assertEqual(textinput._get_selected_text(), '')
+        self.assertEqual(textinput.get_value(), 't')
+
+        # Test readonly
+        textinput.update(PygameEventUtils.key(pygame.K_t, keydown=True, char='k'))
+        self.assertEqual(textinput.get_value(), 'tk')
+        textinput.readonly = True
+        textinput.update(PygameEventUtils.key(pygame.K_t, keydown=True, char='k'))
+        self.assertEqual(textinput.get_value(), 'tk')
+        textinput.readonly = False
+
+        # Test alt+x
+        textinput.update(PygameEventUtils.key(pygame.K_SPACE, keydown=True))
+        textinput.update(PygameEventUtils.key(pygame.K_2, keydown=True, char='2'))
+        textinput.update(PygameEventUtils.key(pygame.K_1, keydown=True, char='1'))
+        textinput.update(PygameEventUtils.key(pygame.K_1, keydown=True, char='5'))
+        self.assertEqual(textinput.get_value(), 'tk 215')
+        textinput.update(PygameEventUtils.keydown_mod_alt(pygame.K_x))  # convert 215 to unicode
+        self.assertEqual(textinput.get_value(), 'tkȕ')
+        textinput.update(PygameEventUtils.key(pygame.K_SPACE, keydown=True))
+        textinput.update(PygameEventUtils.key(pygame.K_SPACE, keydown=True))
+        textinput.update(PygameEventUtils.key(pygame.K_b, keydown=True, char='B'))
+        textinput.update(PygameEventUtils.key(pygame.K_1, keydown=True, char='1'))
+        textinput.update(PygameEventUtils.keydown_mod_alt(pygame.K_x))  # convert 215 to unicode
+        self.assertEqual(textinput.get_value(), 'tkȕ ±')
+
+        # Remove all
+        textinput.clear()
+        textinput.update(PygameEventUtils.key(pygame.K_b, keydown=True, char='B'))
+        textinput.update(PygameEventUtils.key(pygame.K_1, keydown=True, char='1'))
+        textinput.update(PygameEventUtils.keydown_mod_alt(pygame.K_x))  # convert 215 to unicode
+        self.assertEqual(textinput.get_value(), '±')
+        textinput.update(PygameEventUtils.keydown_mod_alt(pygame.K_x))  # convert same to unicode, do nothing
+        self.assertEqual(textinput.get_value(), '±')
+
+        # Test consecutive
+        textinput.update(PygameEventUtils.key(pygame.K_2, keydown=True, char='2'))
+        textinput.update(PygameEventUtils.key(pygame.K_0, keydown=True, char='0'))
+        textinput.update(PygameEventUtils.key(pygame.K_1, keydown=True, char='1'))
+        textinput.update(PygameEventUtils.key(pygame.K_3, keydown=True, char='3'))
+        textinput.update(PygameEventUtils.keydown_mod_alt(pygame.K_x))  # convert 215 to unicode
+        self.assertEqual(textinput.get_value(), '±–')
+
+        # Test 0x
+        textinput.clear()
+        PygameEventUtils.release_key_mod()
+        textinput.update(PygameEventUtils.key(pygame.K_0, keydown=True, char='0'))
+        self.assertEqual(textinput.get_value(), '0')
+        textinput.update(PygameEventUtils.key(pygame.K_x, keydown=True, char='x'))
+        self.assertEqual(textinput.get_value(), '0x')
+        textinput.update(PygameEventUtils.keydown_mod_alt(pygame.K_x))
+        self.assertEqual(textinput.get_value(), '0x')
+        textinput.update(PygameEventUtils.key(pygame.K_b, keydown=True, char='B'))
+        textinput.update(PygameEventUtils.key(pygame.K_1, keydown=True, char='1'))
+        self.assertEqual(textinput.get_value(), '0xB1')
+        textinput.update(PygameEventUtils.keydown_mod_alt(pygame.K_x))
+        self.assertEqual(textinput.get_value(), '±')
+        PygameEventUtils.release_key_mod()
+
+        textinput.update(PygameEventUtils.key(pygame.K_0, keydown=True, char='0'))
+        textinput.update(PygameEventUtils.key(pygame.K_x, keydown=True, char='x'))
+        textinput.update(PygameEventUtils.key(pygame.K_b, keydown=True, char='B'))
+        textinput.update(PygameEventUtils.key(pygame.K_1, keydown=True, char='1'))
+        textinput.update(PygameEventUtils.keydown_mod_alt(pygame.K_x))
+        self.assertEqual(textinput.get_value(), '±±')
+
+        # Test tab
+        self.assertEqual(textinput._tab_size, 4)
+        textinput.update(PygameEventUtils.key(pygame.K_TAB, keydown=True))
+        self.assertEqual(textinput.get_value(), '±±    ')
+
+        # Test mouse
+        textinput._selected = True
+        textinput._selection_time = 0
+        textinput.update(PygameEventUtils.middle_rect_click(textinput))
+        self.assertTrue(textinput._cursor_visible)
+        textinput._select_all()
+        textinput._selection_active = True
+        self.assertEqual(textinput._cursor_position, 6)
+        self.assertEqual(textinput._selection_box, [0, 6])
+        textinput.update(PygameEventUtils.middle_rect_click(textinput, evtype=pygame.MOUSEBUTTONDOWN))
+        self.assertEqual(textinput._selection_box, [0, 0])
+
+        # Check click pos
+        textinput._check_mouse_collide_input(PygameEventUtils.middle_rect_click(textinput)[0].pos)
+        self.assertEqual(textinput._cursor_position, 6)
+
+        # Test touch
+        textinput._cursor_position = 0
+        textinput._check_touch_collide_input(PygameEventUtils.middle_rect_click(textinput)[0].pos)
+        self.assertEqual(textinput._cursor_position, 6)
+
+        # Update mouse
+        for i in range(50):
+            textinput.update(PygameEventUtils.key(pygame.K_t, keydown=True, char='t'))
+        textinput._update_cursor_mouse(50)
+        textinput._cursor_render = True
+        textinput._render_cursor()
+
+        # Test multiple are selected
+        self.assertRaises(pygame_menu.menu._MenuMultipleSelectedWidgetsException, lambda: menu.draw(surface))
+        textinput.clear()
+        textinput.select(update_menu=True)
+        menu.draw(surface)
+
+        # Clear the menu
+        self.assertEqual(menu._stats.removed_widgets, 0)
+        self.assertEqual(textinput.get_menu(), menu)
+        menu.clear()
+        self.assertIsNone(textinput.get_menu())
+        self.assertEqual(menu._stats.removed_widgets, 6)
+        menu.add.generic_widget(textinput)
+        self.assertEqual(textinput.get_menu(), menu)
+        menu.clear()
+        self.assertEqual(menu._stats.removed_widgets, 7)
+
+        # Test text with max width and right overflow removal
+        menu._copy_theme()
+        menu._theme.widget_font_size = 20
+        textinput = menu.add.text_input(
+            'Some long text: ',
+            maxwidth=19,
+            textinput_id='long_text',
+            input_underline='_'
+        )
+        self.assertRaises(WidgetTransformationNotImplemented, lambda: textinput.resize())
+        self.assertRaises(WidgetTransformationNotImplemented, lambda: textinput.set_max_width())
+        self.assertRaises(WidgetTransformationNotImplemented, lambda: textinput.set_max_height())
+        self.assertRaises(WidgetTransformationNotImplemented, lambda: textinput.scale())
+        self.assertRaises(WidgetTransformationNotImplemented, lambda: textinput.rotate())
+        textinput.flip(True, True)
+        self.assertEqual(textinput._flip, (False, True))
+        textinput.set_value('aaaaaaaaaaaaaaaaaaaaaaaaaa')
+        self.assertEqual(textinput._cursor_position, 26)
+        self.assertEqual(textinput._renderbox, [1, 26, 25])
+        textinput.update(PygameEventUtils.key(pygame.K_BACKSPACE, keydown=True))
+        self.assertEqual(textinput._cursor_position, 25)
+        self.assertEqual(textinput._renderbox, [0, 25, 25])
+        textinput.update(PygameEventUtils.key(pygame.K_a, keydown=True, char='a'))
+        self.assertEqual(textinput._cursor_position, 26)
+        self.assertEqual(textinput._renderbox, [1, 26, 25])
+        textinput.update(PygameEventUtils.key(pygame.K_BACKSPACE, keydown=True))
+        self.assertEqual(textinput._cursor_position, 25)
+        self.assertEqual(textinput._renderbox, [0, 25, 25])
+
+        # Test underline edge cases
+        theme = TEST_THEME.copy()
+        theme.title_font_size = 35
+        theme.widget_font_size = 25
+
+        menu = pygame_menu.Menu(
+            column_min_width=400,
+            height=300,
+            theme=theme,
+            title='Label',
+            onclose=pygame_menu.events.CLOSE,
+            width=400
+        )
+        textinput = menu.add.text_input('title', input_underline='_')
+        self.assertEqual(menu._widget_offset[1], 107 if PYGAME_V2 else 106)
+        self.assertEqual(textinput.get_width(), 376)
+        self.assertEqual(textinput._current_underline_string, '______________________________')
+        menu.render()
+        self.assertEqual((menu.get_width(widget=True), menu.get_width(inner=True)), (376, 400))
+        self.assertEqual(textinput.get_width(), 376)
+        self.assertEqual(textinput._current_underline_string, '______________________________')
+        menu.render()
+        self.assertEqual((menu.get_width(widget=True), menu.get_width(inner=True)), (376, 400))
+        textinput.set_title('nice')
+        self.assertEqual(textinput.get_width(), 379)
+        self.assertEqual(textinput._current_underline_string, '______________________________')
+        menu.render()
+        self.assertEqual((menu.get_width(widget=True), menu.get_width(inner=True)), (379, 400))
+        # noinspection SpellCheckingInspection
+        textinput.set_value('QQQQQQQQQQQQQQQQQQQQQQQQQQQQQQQQQQQ')
+        self.assertEqual(textinput.get_width(), 712)
+        self.assertEqual(textinput._current_underline_string,
+                         '____________________________________________________________')
+        menu.render()
+        self.assertEqual((menu.get_width(widget=True), menu.get_width(inner=True)), (712, 400))
+        textinput.set_padding(100)
+        self.assertEqual(textinput.get_width(), 912)
+        self.assertEqual(textinput._current_underline_string,
+                         '____________________________________________________________')
+        menu.render()
+        self.assertEqual((menu.get_width(widget=True), menu.get_width(inner=True)), (912, 380))
+        textinput.set_padding(200)
+        self.assertEqual(textinput.get_width(), 1112)
+        self.assertEqual(textinput._current_underline_string,
+                         '____________________________________________________________')
+        menu.render()
+        self.assertEqual((menu.get_width(widget=True), menu.get_width(inner=True)), (1112, 380))
+
+        # Test underline
+        textinput = menu.add.text_input('title: ')
+        textinput.set_value('this is a test value')
+        self.assertEqual(textinput.get_width(), 266)
+
+        menu.clear()
+        textinput = menu.add.text_input('title: ', input_underline='.-')
+        # noinspection SpellCheckingInspection
+        textinput.set_value('QQQQQQQQQQQQQQQ')
+        self.assertEqual(textinput.get_width(), 373)
+        self.assertEqual(textinput._current_underline_string, '.-.-.-.-.-.-.-.-.-.-.-.-.-.-.-.-.-.-.-.-.-')
+
+        textinput = menu.add.text_input('title: ', input_underline='_', input_underline_len=10)
+        self.assertEqual(textinput._current_underline_string, '_' * 10)
+
+        # Text underline with different column widths
+        menu = pygame_menu.Menu(
+            column_max_width=200,
+            height=300,
+            theme=theme,
+            title='Label',
+            onclose=pygame_menu.events.CLOSE,
+            width=400
+        )
+        self.assertRaises(pygame_menu.menu._MenuSizingException, lambda: menu.add.frame_v(300, 100))
+        self.assertRaises(pygame_menu.menu._MenuSizingException, lambda: menu.add.frame_v(201, 100))
+        self.assertEqual(len(menu._widgets), 0)
+        textinput = menu.add.text_input('title', input_underline='_')
+        self.assertEqual(menu._widget_offset[1], 107 if PYGAME_V2 else 106)
+        self.assertEqual(textinput.get_width(), 178)
+        self.assertEqual(textinput._current_underline_string, '____________')
+        v_frame = menu.add.frame_v(150, 100, background_color=(20, 20, 20))
+        v_frame.pack(textinput)
+        self.assertEqual(menu._widget_offset[1], 76 if PYGAME_V2 else 75)
+        self.assertEqual(textinput.get_width(), 134)
+        self.assertEqual(textinput._current_underline_string, '________')
+
+        # Test cursor size
+        self.assertRaises(AssertionError, lambda: menu.add.text_input('title', cursor_size=(1, 0)))
+        self.assertRaises(AssertionError, lambda: menu.add.text_input('title', cursor_size=(-1, -1)))
+        self.assertRaises(AssertionError, lambda: menu.add.text_input('title', cursor_size=(1, 1, 0)))
+        self.assertRaises(AssertionError, lambda: menu.add.text_input('title', cursor_size=[1, 1]))
+        self.assertRaises(AssertionError, lambda: menu.add.text_input('title', cursor_size=(1.6, 2.5)))
+
+        textinput_cursor = menu.add.text_input('title', cursor_size=(10, 2))
+        self.assertEqual(textinput_cursor._cursor_size, (10, 2))
+
+    # noinspection PyArgumentEqualDefault,PyTypeChecker
+    def test_colorinput(self) -> None:
+        """
+        Test ColorInput widget.
+        """
+        if SYS_PLATFORM_OSX:
+            return
+
+        def _assert_invalid_color(widg) -> None:
+            """
+            Assert that the widget color is invalid.
+            :param widg: Widget object
+            """
+            r, g, b = widg.get_value()
+            self.assertEqual(r, -1)
+            self.assertEqual(g, -1)
+            self.assertEqual(b, -1)
+
+        def _assert_color(widg, cr, cg, cb) -> None:
+            """
+            Assert that the widget color is invalid.
+            :param widg: Widget object
+            :param cr: Red channel, number between 0 and 255
+            :param cg: Green channel, number between 0 and 255
+            :param cb: Blue channel, number between 0 and 255
+            """
+            r, g, b = widg.get_value()
+            self.assertEqual(r, cr)
+            self.assertEqual(g, cg)
+            self.assertEqual(b, cb)
+
+        menu = MenuUtils.generic_menu(theme=TEST_THEME.copy())
+
+        # Base rgb
+        widget = menu.add.color_input('title', color_type='rgb', input_separator=',')
+        widget.set_value((123, 234, 55))
+        self.assertRaises(AssertionError,
+                          lambda: widget.set_value('0,0,0'))
+        self.assertRaises(AssertionError,
+                          lambda: widget.set_value((255, 0,)))
+        self.assertRaises(AssertionError,
+                          lambda: widget.set_value((255, 255, -255)))
+        _assert_color(widget, 123, 234, 55)
+
+        # Test separator
+        widget = menu.add.color_input('color', color_type='rgb', input_separator='+')
+        widget.set_value((34, 12, 12))
+        self.assertEqual(widget._input_string, '34+12+12')
+        self.assertRaises(AssertionError,
+                          lambda: menu.add.color_input('title', color_type='rgb', input_separator=''))
+        self.assertRaises(AssertionError,
+                          lambda: menu.add.color_input('title', color_type='rgb', input_separator='  '))
+        self.assertRaises(AssertionError,
+                          lambda: menu.add.color_input('title', color_type='unknown'))
+        for i in range(10):
+            self.assertRaises(AssertionError,
+                              lambda: menu.add.color_input('title', color_type='rgb', input_separator=str(i)))
+
+        # Empty rgb
+        widget = menu.add.color_input('color', color_type='rgb', input_separator=',')
+
+        PygameEventUtils.test_widget_key_press(widget)
+        self.assertEqual(widget._cursor_position, 0)
+        widget.update(PygameEventUtils.key(pygame.K_RIGHT, keydown=True))
+        self.assertEqual(widget._cursor_position, 0)
+        _assert_invalid_color(widget)
+
+        # Write sequence: 2 -> 25 -> 25, -> 25,0,
+        # The comma after the zero must be automatically set
+        self.assertFalse(widget.update(PygameEventUtils.key(0, keydown=True, testmode=False)))
+        widget.update(PygameEventUtils.key(pygame.K_2, keydown=True, char='2'))
+        widget.update(PygameEventUtils.key(pygame.K_5, keydown=True, char='5'))
+        widget.update(PygameEventUtils.key(pygame.K_COMMA, keydown=True, char=','))
+        self.assertEqual(widget._input_string, '25,')
+        widget.update(PygameEventUtils.key(pygame.K_0, keydown=True, char='0'))
+        self.assertEqual(widget._input_string, '25,0,')
+        _assert_invalid_color(widget)
+
+        # Now, sequence: 25,0,c -> 25c,0, with cursor c
+        widget.update(PygameEventUtils.key(pygame.K_LEFT, keydown=True))
+        widget.update(PygameEventUtils.key(pygame.K_LEFT, keydown=True))
+        widget.update(PygameEventUtils.key(pygame.K_LEFT, keydown=True))
+        self.assertEqual(widget._cursor_position, 2)
+
+        # Sequence. 25,0, -> 255,0, -> 255,0, trying to write another 5 in the same position
+        # That should be cancelled because 2555 > 255
+        widget.update(PygameEventUtils.key(pygame.K_5, keydown=True, char='5'))
+        self.assertEqual(widget._input_string, '255,0,')
+        widget.update(PygameEventUtils.key(pygame.K_5, keydown=True, char='5'))
+        self.assertEqual(widget._input_string, '255,0,')
+
+        # Invalid left zeros, try to write 255,0, -> 255,00, but that should be disabled
+        widget.update(PygameEventUtils.key(pygame.K_RIGHT, keydown=True))
+        widget.update(PygameEventUtils.key(pygame.K_0, keydown=True, char='0'))
+        self.assertEqual(widget._input_string, '255,0,')
+
+        # Second comma cannot be deleted because there's a number between ,0,
+        widget.update(PygameEventUtils.key(pygame.K_BACKSPACE, keydown=True))
+        self.assertEqual(widget._input_string, '255,0,')
+        widget.update(PygameEventUtils.key(pygame.K_LEFT, keydown=True))
+        widget.update(PygameEventUtils.key(pygame.K_DELETE, keydown=True))
+        self.assertEqual(widget._input_string, '255,0,')
+
+        # Current cursor is at 255c,0,
+        # Now right comma and 0 can be deleted
+        widget.update(PygameEventUtils.key(pygame.K_END, keydown=True))
+        widget.update(PygameEventUtils.key(pygame.K_BACKSPACE, keydown=True))
+        widget.update(PygameEventUtils.key(pygame.K_BACKSPACE, keydown=True))
+        self.assertEqual(widget._input_string, '255,')
+
+        # Fill with zeros, then number with 2 consecutive 0 types must be 255,0,0
+        # Commas should be inserted automatically
+        widget.readonly = True
+        widget.update(PygameEventUtils.key(pygame.K_0, keydown=True, char='0'))
+        self.assertEqual(widget._input_string, '255,')
+        widget.readonly = False
+        widget.update(PygameEventUtils.key(pygame.K_0, keydown=True, char='0'))
+        widget.update(PygameEventUtils.key(pygame.K_0, keydown=True, char='0'))
+        self.assertEqual(widget._input_string, '255,0,0')
+        _assert_color(widget, 255, 0, 0)
+
+        # At this state, user cannot add more zeros at right
+        for i in range(5):
+            widget.update(PygameEventUtils.key(pygame.K_0, keydown=True, char='0'))
+        self.assertEqual(widget._input_string, '255,0,0')
+        widget.get_rect()
+
+        widget.clear()
+        self.assertEqual(widget._input_string, '')
+
+        # Assert invalid defaults rgb
+        self.assertRaises(AssertionError,
+                          lambda: menu.add.color_input('title', color_type='rgb', default=(255, 255,)))
+        self.assertRaises(AssertionError,
+                          lambda: menu.add.color_input('title', color_type='rgb', default=(255, 255)))
+        self.assertRaises(AssertionError,
+                          lambda: menu.add.color_input('title', color_type='rgb', default=(255, 255, 255, 255)))
+
+        # Assert hex widget
+        widget = menu.add.color_input('title', color_type='hex')
+        self.assertEqual(widget._input_string, '#')
+        self.assertEqual(widget._cursor_position, 1)
+        _assert_invalid_color(widget)
+        self.assertRaises(AssertionError,
+                          lambda: widget.set_value('#FF'))
+        self.assertRaises(AssertionError,
+                          lambda: widget.set_value('#FFFFF<'))
+        self.assertRaises(AssertionError,
+                          lambda: widget.set_value('#FFFFF'))
+        self.assertRaises(AssertionError,
+                          lambda: widget.set_value('#F'))
+        # noinspection SpellCheckingInspection
+        self.assertRaises(AssertionError,
+                          lambda: widget.set_value('FFFFF'))
+        self.assertRaises(AssertionError,
+                          lambda: widget.set_value('F'))
+        widget.set_value('FF00FF')
+        _assert_color(widget, 255, 0, 255)
+        widget.set_value('#12FfAa')
+        _assert_color(widget, 18, 255, 170)
+        widget.set_value('   59C1e5')
+        _assert_color(widget, 89, 193, 229)
+
+        widget.render()
+        widget.draw(surface)
+
+        widget.clear()
+        self.assertEqual(widget._input_string, '#')  # This cannot be empty
+        self.assertEqual(widget._cursor_position, 1)
+
+        # In hex widget # cannot be deleted
+        widget.update(PygameEventUtils.key(pygame.K_BACKSPACE, keydown=True))
+        self.assertEqual(widget._cursor_position, 1)
+        widget.update(PygameEventUtils.key(pygame.K_LEFT, keydown=True))
+        widget.update(PygameEventUtils.key(pygame.K_DELETE, keydown=True))
+        self.assertEqual(widget._input_string, '#')
+        widget.update(PygameEventUtils.key(pygame.K_END, keydown=True))
+        for i in range(10):
+            widget.update(PygameEventUtils.key(pygame.K_f, keydown=True, char='f'))
+        self.assertEqual(widget._input_string, '#ffffff')
+        _assert_color(widget, 255, 255, 255)
+
+        # Test hex formats
+        widget = menu.add.color_input('title', color_type='hex', hex_format='none')
+        widget.set_value('#ff00ff')
+        self.assertFalse(widget.update(PygameEventUtils.key(0, keydown=True, testmode=False)))
+        self.assertEqual(widget.get_value(as_string=True), '#ff00ff')
+        widget.set_value('#FF00ff')
+        self.assertEqual(widget.get_value(as_string=True), '#FF00ff')
+
+        widget = menu.add.color_input('title', color_type='hex', hex_format='lower')
+        widget.set_value('#FF00ff')
+        self.assertEqual(widget.get_value(as_string=True), '#ff00ff')
+        widget.set_value('AABBcc')
+        self.assertEqual(widget.get_value(as_string=True), '#aabbcc')
+
+        widget = menu.add.color_input('title', color_type='hex', hex_format='upper')
+        widget.set_value('#FF00ff')
+        self.assertEqual(widget.get_value(as_string=True), '#FF00FF')
+        widget.set_value('AABBcc')
+        self.assertEqual(widget.get_value(as_string=True), '#AABBCC')
+
+        # Test dynamic sizing
+        widget = menu.add.color_input('title', color_type='hex', hex_format='upper', dynamic_width=True)
+        self.assertEqual(widget.get_width(), 200)
+        widget.set_value('#ffffff')
+        width = 342 if PYGAME_V2 else 345
+        self.assertEqual(widget.get_width(), width)
+        widget.set_value(None)
+        self.assertEqual(widget.get_width(), 200)
+        self.assertEqual(widget.get_value(as_string=True), '#')
+        widget.set_value('#ffffff')
+        self.assertEqual(widget.get_width(), width)
+        widget.update(
+            PygameEventUtils.key(pygame.K_BACKSPACE,
+                                 keydown=True))  # remove the last character, now color is invalid
+        self.assertEqual(widget.get_value(as_string=True), '#FFFFF')  # is upper
+        widget.render()
+        self.assertEqual(widget.get_width(), 200)
+
+        widget = menu.add.color_input('title', color_type='hex', hex_format='upper', dynamic_width=False)
+        self.assertEqual(widget.get_width(), width)
+        widget.set_value('#ffffff')
+        self.assertEqual(widget.get_width(), width)
